dev_requirements.txt
django>=1.11
libsass
watchdog